chore: rename mistyped file
